cloudfunction/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery
google-cloud-storage